Change in a R Source of a ComponentList does not refresh the controls correctly
In the Cochrane Meta Analysis, a ComponentsList of a ComponentsLists of Checkboxes R Sources. These R sources are changed according to some options.
The ComponentsList were not updated correctly each time the R sources changed.
This was due to bad caching (the _rSourceModelMap in AnalysisForm was not updated correctly and is in fact not needed), and by not signaling that the analysis should be rerun again.